requirements.txt
dj-database-url==0.4.1
Django==1.10.3
django-bootstrap3==7.0.1
django-configurations==2.0
django-tables2==1.2.6
gunicorn==19.6.0
psycopg2==2.6.2
whitenoise==3.2.2
